+++ app/src/main/AndroidManifest.xml
@@ -30,6 +30,8 @@
         <activity android:name=".RoomActivity"/>
         <activity android:name=".StatsActivity"/>
         <activity android:name=".ProfileActivity"/>
+        <activity android:name=".RoomListActivity"/>
+        <activity android:name=".SetPasswordActivity"/>
     </application>
 
 </manifest>
